[GUI] Refactor updateStakingStatus and set it to inactive after locking
Directly update the staking status icon when the wallet is manually
locked so the user doesn't have to wait for the polling timeout and gets
instant feedback.

Otherwise the wallet would have (even if just for 15-30 seconds)
simultaneously the staking icon ON and the lock icon LOCKED, which could
be quite confusing.